+++ presentation/src/main/java/com/threegap/bitnagil/presentation/splash/SplashScreen.kt
@@ -25,7 +25,7 @@ import com.threegap.bitnagil.designsystem.component.atom.BitnagilIcon
 import com.threegap.bitnagil.presentation.common.playstore.openAppInPlayStore
 import com.threegap.bitnagil.presentation.splash.component.template.BitnagilLottieAnimation
 import com.threegap.bitnagil.presentation.splash.component.template.ForceUpdateDialog
-import com.threegap.bitnagil.presentation.splash.model.SplashSideEffect
+import com.threegap.bitnagil.presentation.splash.contract.SplashSideEffect
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
 import kotlin.system.exitProcess
